py/requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
psycopg2-binary
pandas
fpdf
matplotlib
fastapi
uvicorn
python-multipart
pydantic
